requirements.txt
flask
flask-login
requests
google-cloud-firestore
google-cloud-storage
pyopenssl
google-cloud-vision
